Direct users who add non-english translation changes to crowdin
This commit adds a new script that will comment on a pull request if it is opened with non-English translation files changed, directing the user to the crowdin project. Also added a new function that returns the list of files changed in the given PR number. Tests were included to test the new functionality.

Signed-off-by: Yarden Shoham <git@yardenshoham.com>