Lock issues closed for 3 months and pull requests merged/closed for 3 months automatically
We will lock issues and pull requests that have been closed for 3 months. If the issue was updated in the last month, a comment will be posted suggesting opening a new issue to continue the discussion.

The check is done on every push and will lock at most 30 issues at a time. We currently have ~400 issues waiting locking so the initial locking will happen over the next few days.

Signed-off-by: Yarden Shoham <git@yardenshoham.com>